requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
colorama==0.4.3
cryptography==35.0.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.2.8
django-allauth==0.42.0
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-crispy-forms==1.9.2
django-environ==0.8.1
django-etc==1.3.0
django-extensions==3.0.9
django-hitcount==1.3.3
django-js-asset==1.2.2
django-markdownx==3.0.1
django-social-share==2.2.1
django-summernote==0.8.20.0
environ==1.0
gunicorn==20.1.0
idna==2.10
image==1.5.33
importlib-metadata==2.0.0
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kakaocert==2.50.0
linkhub==1.4.0
Markdown==3.2.2
oauthlib==3.1.0
packaging==21.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.7
psycopg-binary==3.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.2.0
